# api/openapi.yaml
openapi: 3.1.0
info:
  title: AI-BOS API
  version: 0.2.0
  description: |
    Core REST API contract for the AI Business Operating System (AI-BOS).

    Clients may send `X-Request-Id`. The API preserves valid values (letters,
    digits, '.', '_', and '-' up to 128 characters) or generates a unique ID.
    Every response returns the effective ID in the `X-Request-Id` header and
    in the response JSON envelope.
servers:
  - url: /
paths:
...
  /api/v1/auth/register:
    post:
      operationId: register
      summary: Register a new user account
      description: Registers a user with email and password, creates an optional initial business, and returns auth tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/auth/login:
    post:
      operationId: login
      summary: Authenticate with email and password
      description: Validates user credentials and issues an access token along with a rotating refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication successful.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v1/auth/refresh:
    post:
      operationId: refreshTokens
      summary: Refresh access token
      description: Exchanges a valid rotating refresh token (from cookie or payload) for a new access token and rotated refresh token.
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Tokens refreshed successfully.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke active session
      description: Revokes the current refresh session and clears session cookies.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logged out successfully.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessageResponse"
  /api/v1/auth/verify-email:
    post:
      operationId: verifyEmail
      summary: Verify user email address
      description: Consumes an email verification token and marks the account as verified.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
      responses:
        "200":
          description: Email verified successfully.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessageResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
  /api/v1/auth/forgot-password:
    post:
      operationId: forgotPassword
      summary: Request password reset link
      description: Sends a password reset token to the registered email address if found.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Password reset link dispatched if account exists.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessageResponse"
  /api/v1/auth/reset-password:
    post:
      operationId: resetPassword
      summary: Reset password with token
      description: Updates password using a valid reset token and revokes all active sessions.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset successfully.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessageResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
  /api/v1/me:
    get:
      operationId: getCurrentUser
      summary: Get current authenticated user profile and memberships
      description: Returns the authenticated user record and all businesses where the user has active membership.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Authenticated user details.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/v1/businesses:
    get:
      operationId: listBusinesses
      summary: List user businesses
      description: Returns all businesses the authenticated user belongs to.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of accessible businesses.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      operationId: createBusiness
      summary: Create a new business tenant
      description: Creates a new business tenant and assigns the creator as owner.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessRequest"
      responses:
        "201":
          description: Business tenant created.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessDetailResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained via /api/v1/auth/login or /api/v1/auth/register.
  headers:
    RequestId:
      description: Correlates the response with API logs and traces.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^[A-Za-z0-9._-]+$"
...
  responses:
    BadRequestError:
      description: Malformed request or unparseable input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnauthorizedError:
      description: Authentication is missing, expired, or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ConflictError:
      description: State conflict, such as email already registered or duplicate membership.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Semantic validation failed on one or more request fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ResponseMeta:
      type: object
      additionalProperties: false
      required:
        - requestId
      properties:
        requestId:
          type: string
          minLength: 1
          maxLength: 128
    StandardSuccessResponse:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data: {}
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    ErrorDetail:
      type: object
      additionalProperties: true
    StandardError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
          description: Human-safe message that does not expose internal errors or secrets.
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        requestId:
          type: string
          minLength: 1
          maxLength: 128
        details:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/StandardError"
    MessageData:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
    SuccessMessageResponse:
      allOf:
        - $ref: "#/components/schemas/StandardSuccessResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/MessageData"
    RegisterRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 120
        businessName:
          type: string
          minLength: 1
          maxLength: 120
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      additionalProperties: false
      properties:
        refreshToken:
          type: string
    VerifyEmailRequest:
      type: object
      additionalProperties: false
      required:
        - token
      properties:
        token:
          type: string
          minLength: 1
    ForgotPasswordRequest:
      type: object
      additionalProperties: false
      required:
        - email
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      additionalProperties: false
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 8
          maxLength: 128
    UserData:
      type: object
      additionalProperties: false
      required:
        - id
        - email
        - name
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        status:
          type: string
          enum: [active, unverified, suspended]
        verifiedAt:
          type:
            - string
            - "null"
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AuthTokensData:
      type: object
      additionalProperties: false
      required:
        - accessToken
        - tokenType
        - expiresIn
        - user
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
          description: Included in JSON when cookie transport is not active.
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          description: Expiry duration in seconds.
          example: 900
        user:
          $ref: "#/components/schemas/UserData"
    AuthResponse:
      allOf:
        - $ref: "#/components/schemas/StandardSuccessResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/AuthTokensData"
    UserMembershipSummary:
      type: object
      additionalProperties: false
      required:
        - businessId
        - businessName
        - role
        - status
      properties:
        businessId:
          type: string
          format: uuid
        businessName:
          type: string
        role:
          type: string
          enum: [owner, staff, admin]
        status:
          type: string
          enum: [active, invited, suspended]
    MeData:
      type: object
      additionalProperties: false
      required:
        - user
        - memberships
      properties:
        user:
          $ref: "#/components/schemas/UserData"
        memberships:
          type: array
          items:
            $ref: "#/components/schemas/UserMembershipSummary"
    MeResponse:
      allOf:
        - $ref: "#/components/schemas/StandardSuccessResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/MeData"
    BusinessData:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - timezone
        - currency
        - invoicePrefix
        - invoiceSequence
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        legalName:
          type:
            - string
            - "null"
        phone:
          type:
            - string
            - "null"
        address:
          type:
            - string
            - "null"
        timezone:
          type: string
          example: Asia/Kolkata
        currency:
          type: string
          example: INR
        invoicePrefix:
          type: string
          example: INV-
        invoiceSequence:
          type: integer
          example: 1001
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateBusinessRequest:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        legalName:
          type: string
          maxLength: 150
        phone:
          type: string
          maxLength: 30
        address:
          type: string
          maxLength: 500
        timezone:
          type: string
          maxLength: 50
          default: Asia/Kolkata
        currency:
          type: string
          maxLength: 3
          default: INR
        invoicePrefix:
          type: string
          maxLength: 10
          default: INV-
...
    BusinessDetailResponse:
      allOf:
        - $ref: "#/components/schemas/StandardSuccessResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/BusinessData"
    BusinessListResponse:
      allOf:
        - $ref: "#/components/schemas/StandardSuccessResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/BusinessData"
...